+++ nucypher/cli/characters/bob.py
