+++ nucypher/cli/main.py
@@ -19,9 +19,10 @@ along with nucypher.  If not, see <https://www.gnu.org/licenses/>.
 
 import click
 
-from nucypher.cli import moe, ursula, status, alice, bob, enrico
+from nucypher.characters.banners import BANNER
+from nucypher.cli import status
+from nucypher.cli.characters import moe, bob, ursula, alice, enrico
 from nucypher.cli.config import echo_version, nucypher_click_config
-from nucypher.cli.painting import BANNER
 from nucypher.utilities.logging import GlobalConsoleLogger
 
 GlobalConsoleLogger.start_if_not_started()
